+++ scripts/ts/ticket/jira/index.ts
